add first version of layer ensembling to eval
